requirements.txt
# FastAPI MVC Backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
email-validator>=2.0.0

# SHL Recommendation Engine (scraper, retrieval, GenAI)
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
rank_bm25>=0.2.2
google-generativeai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
pandas>=2.1.0
openpyxl>=3.1.0
playwright>=1.40.0

# Dev
mypy>=1.8.0
ruff>=0.1.0
pytest>=7.4.0
